moving function to utils to stop circular imports
