+++ jwql/website/apps/jwql/data_containers.py
@@ -53,6 +53,7 @@ from jwql.utils.constants import IGNORED_SUFFIXES, INSTRUMENT_SERVICE_MATCH, JWS
                                  JWST_INSTRUMENT_NAMES_SHORTHAND, SUFFIXES_TO_ADD_ASSOCIATION, SUFFIXES_WITH_AVERAGED_INTS
 from jwql.utils.preview_image import PreviewImage
 from jwql.utils.credentials import get_mast_token
+from jwql.utils.utils import get_rootnames_for_instrument_proposal
 from .forms import InstrumentAnomalySubmitForm
 from astroquery.mast import Mast
 from jwedb.edb_interface import mnemonic_inventory
@@ -1032,30 +1033,6 @@ def get_proposal_info(filepaths):
     proposal_info['observation_nums'] = observations
 
     return proposal_info
-
-
-def get_rootnames_for_instrument_proposal(instrument, proposal):
-    """Return a list of rootnames for the given instrument and proposal
-
-    Parameters
-    ----------
-    instrument : str
-        Name of the JWST instrument, with first letter capitalized
-        (e.g. ``Fgs``)
-
-    proposal : int or str
-        Proposal ID number
-
-    Returns
-    -------
-    rootnames : list
-        List of rootnames for the given instrument and proposal number
-    """
-    tap_service = vo.dal.TAPService("http://vao.stsci.edu/caomtap/tapservice.aspx")
-    tap_results = tap_service.search(f"select observationID from dbo.CaomObservation where collection='JWST' and maxLevel=2 and insName like '{instrument.lower()}' and prpID='{int(proposal)}'")
-    prop_table = tap_results.to_table()
-    rootnames = prop_table['observationID'].data
-    return rootnames.compressed()
 
 
 def get_rootnames_for_proposal(proposal):
